Allow a goal component to be displayed in BaseGoalScreen
