+++ src/components/GoalTimeline/GoalHistory/tests/GoalHistoryComponent.test.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { act } from "react-dom/test-utils";
 import configureMockStore from "redux-mock-store";
+import GoalHistory from "../../GoalHistory";
 import { defaultState } from "../../DefaultState";
 import { Provider } from "react-redux";
-import GoalSwitcher from "..";
 
 const createMockStore = configureMockStore([]);
 
@@ -14,15 +13,12 @@ it("renders without crashing", () => {
       ...defaultState
     }
   });
-  const container = document.createElement("div");
-  act(() => {
-    ReactDOM.render(
-      <Provider store={mockStore}>
-        <GoalSwitcher />
-      </Provider>,
-      container
-    );
-  });
-
-  ReactDOM.unmountComponentAtNode(container);
+  const div = document.createElement("div");
+  ReactDOM.render(
+    <Provider store={mockStore}>
+      <GoalHistory />
+    </Provider>,
+    div
+  );
+  ReactDOM.unmountComponentAtNode(div);
 });
