+++ src/components/GoalTimeline/GoalSuggestions/tests/GoalSuggestions.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import configureMockStore from "redux-mock-store";
-import GoalHistory from "../";
+import GoalSuggestions from "../../GoalSuggestions";
 import { defaultState } from "../../DefaultState";
 import { Provider } from "react-redux";
 
@@ -16,7 +16,7 @@ it("renders without crashing", () => {
   const div = document.createElement("div");
   ReactDOM.render(
     <Provider store={mockStore}>
-      <GoalHistory />
+      <GoalSuggestions />
     </Provider>,
     div
   );
