+++ src/rootReducer.tsx
@@ -2,7 +2,7 @@ import { combineReducers, Reducer } from "redux";
 
 import { StoreState } from "./types";
 import { tempReducer } from "./components/Temp/TempReducer";
-import { goalsReducer } from "./components/GoalView/GoalTimelineReducers";
+import { goalsReducer } from "./components/GoalTimeline/GoalTimelineReducers";
 import { navReducer } from "./components/Navigation/NavigationReducer";
 import { localizeReducer } from "react-localize-redux";
 import { dragWordReducer } from "./goals/DraggableWord/reducer";
